requirements.txt
absl-py==0.8.1
astor==0.8.0
atari-py==0.2.6
certifi==2019.9.11
chardet==3.0.4
cloudpickle==1.2.2
cycler==0.10.0
DataProperty==0.43.1
decorator==4.4.0
future==0.17.1
gast==0.2.2
gin-config==0.2.1
google-pasta==0.1.7
grpcio==1.24.1
gym==0.14.0
h5py==2.10.0
idna==2.8
imageio==2.5.0
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mbstrdecoder==0.8.1
msgfy==0.0.7
networkx==2.3
numpy==1.16.5
opencv-python==4.1.1.26
opt-einsum==3.1.0
pandas==0.25.1
pathvalidate==0.29.0
Pillow==6.2.0
progressbar2==3.39.3
protobuf==3.10.0
pybullet==2.5.6
-e git+https://github.com/benelot/pybullet-gym.git@55eaa0defca7f4ae382963885a334c952133829d#egg=pybulletgym
pyglet==1.3.2
pyparsing==2.4.2
pytablewriter==0.46.1
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.3
PyWavelets==1.0.3
requests==2.21.0
scikit-image==0.15.0
scikit-learn==0.21.0
scipy==1.2.1
six==1.12.0
tabledata==0.9.1
tensorboard==2.0.0
tensorflow-estimator==2.0.0
tensorflow-gpu==2.0.0
tensorflow-probability==0.8.0
tensorlayer==2.1.0
termcolor==1.1.0
typepy==0.6.0
urllib3==1.24.3
Werkzeug==0.16.0
wrapt==1.11.1
